fix:[이채원] 엔티티 ípk 타입 변경 및 이메일 삭제

+++ src/main/java/com/merge/final_project/user/verify/EmailVerification.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class EmailVerification {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer emailVerifyNo;
+    private Long emailVerifyNo;
 
     @Column(name="email")
     private String email;
